+++ internal/conn/conn.go
@@ -46,34 +46,57 @@ func (c *conn) Address() string {
 }
 
 type conn struct {
-	sync.RWMutex
-	config    Config // ro access
-	cc        *grpc.ClientConn
-	done      chan struct{}
-	endpoint  endpoint.Endpoint // ro access
-	closed    bool
-	state     State
-	usages    int32
-	lastUsage time.Time
-	onClose   []func(*conn)
+	mtx          sync.RWMutex
+	config       Config // ro access
+	cc           *grpc.ClientConn
+	done         chan struct{}
+	endpoint     endpoint.Endpoint // ro access
+	closed       bool
+	state        State
+	usages       int32
+	streamUsages int32
+	lastUsage    time.Time
+	onClose      []func(*conn)
 }
 
 func (c *conn) Release(ctx context.Context) {
-	if c.decUsages() == 0 {
-		_ = c.Close(ctx)
+	var (
+		onDone = trace.DriverOnConnRelease(
+			c.config.Trace(),
+			&ctx,
+			c.endpoint.Copy(),
+		)
+		err error
+	)
+	defer func() {
+		onDone(err)
+	}()
+	var issues []error
+	if c.changeUsages(-1) == 0 {
+		if usages := atomic.LoadInt32(&c.streamUsages); usages > 0 {
+			issues = append(issues, fmt.Errorf("conn in stream use: usages=%d", usages))
+		}
+		if closeErr := c.Close(ctx); closeErr != nil {
+			issues = append(issues, closeErr)
+		}
+	}
+	if len(issues) > 0 {
+		err = errors.NewWithIssues("conn released with issues", issues...)
 	}
 }
 
 func (c *conn) LastUsage() time.Time {
-	c.RLock()
-	defer c.RUnlock()
-
+	if usages := atomic.LoadInt32(&c.streamUsages); usages > 0 {
+		return time.Now()
+	}
+	c.mtx.RLock()
+	defer c.mtx.RUnlock()
 	return c.lastUsage
 }
 
 func (c *conn) IsState(states ...State) bool {
-	c.RLock()
-	defer c.RUnlock()
+	c.mtx.RLock()
+	defer c.mtx.RUnlock()
 
 	for _, s := range states {
 		if s == c.state {
@@ -85,8 +108,8 @@ func (c *conn) IsState(states ...State) bool {
 }
 
 func (c *conn) park(ctx context.Context) (err error) {
-	c.Lock()
-	defer c.Unlock()
+	c.mtx.Lock()
+	defer c.mtx.Unlock()
 
 	if c.closed {
 		return nil
@@ -129,8 +152,8 @@ func (c *conn) Endpoint() endpoint.Endpoint {
 }
 
 func (c *conn) SetState(s State) State {
-	c.Lock()
-	defer c.Unlock()
+	c.mtx.Lock()
+	defer c.mtx.Unlock()
 	return c.setState(s)
 }
 
@@ -145,8 +168,8 @@ func (c *conn) setState(s State) State {
 }
 
 func (c *conn) GetState() (s State) {
-	c.RLock()
-	defer c.RUnlock()
+	c.mtx.RLock()
+	defer c.mtx.RUnlock()
 	return c.state
 }
 
@@ -168,8 +191,8 @@ func (c *conn) take(ctx context.Context) (cc *grpc.ClientConn, err error) {
 		)
 	}
 
-	c.Lock()
-	defer c.Unlock()
+	c.mtx.Lock()
+	defer c.mtx.Unlock()
 
 	if !isBroken(c.cc) {
 		return c.cc, nil
@@ -194,30 +217,44 @@ func (c *conn) take(ctx context.Context) (cc *grpc.ClientConn, err error) {
 	return c.cc, nil
 }
 
-func (c *conn) changeUsages(delta int32) int32 {
-	if usages := atomic.AddInt32(&c.usages, delta); usages < 0 {
-		panic("negative usages" + strconv.Itoa(int(usages)))
-	} else {
-		trace.DriverOnConnUsagesChange(
-			c.config.Trace(),
-			c.endpoint.Copy(),
-			int(usages),
-		)
-		return usages
-	}
-}
-
-func (c *conn) incUsages() {
-	c.Lock()
-	defer c.Unlock()
+func (c *conn) touchLastUsage() {
+	c.mtx.Lock()
+	defer c.mtx.Unlock()
 	c.lastUsage = time.Now()
-	c.changeUsages(1)
 }
 
-func (c *conn) decUsages() int32 {
-	c.Lock()
-	defer c.Unlock()
-	return c.changeUsages(-1)
+func (c *conn) changeUsages(delta int32) int32 {
+	defer c.touchLastUsage()
+
+	usages := atomic.AddInt32(&c.usages, delta)
+
+	if usages < 0 {
+		panic("negative usages" + strconv.Itoa(int(usages)))
+	}
+
+	trace.DriverOnConnUsagesChange(
+		c.config.Trace(),
+		c.endpoint.Copy(),
+		int(usages),
+	)
+
+	return usages
+}
+
+func (c *conn) changeStreamUsages(delta int32) {
+	defer c.touchLastUsage()
+
+	usages := atomic.AddInt32(&c.streamUsages, delta)
+
+	if usages < 0 {
+		panic("negative stream usages" + strconv.Itoa(int(usages)))
+	}
+
+	trace.DriverOnConnStreamUsagesChange(
+		c.config.Trace(),
+		c.endpoint.Copy(),
+		int(usages),
+	)
 }
 
 func isBroken(raw *grpc.ClientConn) bool {
@@ -240,14 +277,14 @@ func (c *conn) close() (err error) {
 }
 
 func (c *conn) isClosed() bool {
-	c.RLock()
-	defer c.RUnlock()
+	c.mtx.RLock()
+	defer c.mtx.RUnlock()
 	return c.closed
 }
 
 func (c *conn) Close(ctx context.Context) (err error) {
-	c.Lock()
-	defer c.Unlock()
+	c.mtx.Lock()
+	defer c.mtx.Unlock()
 
 	if c.closed {
 		return nil
@@ -292,8 +329,17 @@ func (c *conn) invoke(
 		)
 	}
 
-	c.incUsages()
-	defer c.decUsages()
+	ctx, err = c.config.Meta().Meta(ctx)
+	if err != nil {
+		return errors.NewGrpcError(
+			codes.Unavailable,
+			errors.WithMsg("ydb driver conn apply meta failed"),
+			errors.WithErr(err),
+		)
+	}
+
+	c.changeUsages(1)
+	defer c.changeUsages(-1)
 
 	return cc.Invoke(ctx, method, req, res, opts...)
 }
@@ -370,8 +416,17 @@ func (c *conn) newStream(
 		)
 	}
 
-	c.incUsages()
-	defer c.decUsages()
+	ctx, err = c.config.Meta().Meta(ctx)
+	if err != nil {
+		return nil, errors.NewGrpcError(
+			codes.Unavailable,
+			errors.WithMsg("ydb driver conn apply meta failed"),
+			errors.WithErr(err),
+		)
+	}
+
+	c.changeStreamUsages(1)
+	defer c.changeStreamUsages(-1)
 
 	return cc.NewStream(ctx, desc, method, opts...)
 }