+++ internal/conn/grpc_client_stream.go
@@ -47,8 +47,8 @@ func (s *grpcClientStream) Context() context.Context {
 }
 
 func (s *grpcClientStream) SendMsg(m interface{}) (err error) {
-	s.c.incUsages()
-	defer s.c.decUsages()
+	s.c.changeStreamUsages(1)
+	defer s.c.changeStreamUsages(-1)
 
 	err = s.s.SendMsg(m)
 	if err != nil && s.wrapping {
@@ -64,8 +64,8 @@ func (s *grpcClientStream) SendMsg(m interface{}) (err error) {
 }
 
 func (s *grpcClientStream) RecvMsg(m interface{}) (err error) {
-	s.c.incUsages()
-	defer s.c.decUsages()
+	s.c.changeStreamUsages(1)
+	defer s.c.changeStreamUsages(-1)
 
 	defer func() {
 		onDone := s.recv(errors.HideEOF(err))